you-play-app/auth-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pydantic
python-jose